core/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
psycopg[binary]==3.2.1
black==24.8.0
ruff==0.6.3
django-cors-headers==4.4.0
djoser==2.2.3
drf-nested-routers==0.93.4
drf-spectacular==0.27.1
django-environ==0.11.2
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
Faker==22.2.0
